[refactor]: abstract pagination function into a separate file
