+++ server/services/AdhocModelService.js
@@ -230,7 +230,6 @@ class AdhocModelService {
       });
     })
     .catch((err) => {
-      console.log(err);
       throw err;
     });
   }
